services/user-service/app/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pyodbc==5.0.1
pydantic==2.5.0
pydantic-settings==2.1.0
kafka-python==2.0.2
bcrypt==4.1.2
passlib==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6